[SU.2] Fix some -Weverything

+++ src/su/cs-ctype.c
@@ -38,7 +38,7 @@ su_cs_cmp_case(char const *cp1, char const *cp2){
    ASSERT_NYD_EXEC(cp2 != NIL, rv = 1);
 
    for(;;){
-      u8 c1, c2;
+      s32 c1, c2;
 
       c1 = su_cs_to_lower(*cp1++);
       c2 = su_cs_to_lower(*cp2++);
@@ -58,7 +58,7 @@ su_cs_cmp_case_n(char const *cp1, char const *cp2, uz n){
    ASSERT_NYD_EXEC(cp2 != NIL, rv = 1);
 
    for(rv = 0; n != 0; --n){
-      u8 c1, c2;
+      s32 c1, c2;
 
       c1 = su_cs_to_lower(*cp1++);
       c2 = su_cs_to_lower(*cp2++);
